pyproject.toml
[tool.poetry]
name = "log-analyzer-with-ui"
version = "0.1.0"
description = ""
authors = ["Talha Eryasar <talha.eryasar@virginmediao2.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
python-evtx = "^0.7.4"
matplotlib = "^3.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
